Storage packing for RewardManager

+++ contracts/interfaces/IRewardManager.sol
@@ -2,7 +2,7 @@
 pragma solidity 0.8.9;
 
 interface IRewardManager {
-    function userRewardAccrued(address token, address user) external view returns (uint256);
+    function userRewardAccrued(address token, address user) external view returns (uint128);
 
-    function rewardState(address token) external view returns (uint256 index, uint256 lastBalance);
+    function rewardState(address token) external view returns (uint128 index, uint128 lastBalance);
 }
